+++ src/lib/index.ts
@@ -8,7 +8,7 @@ export { default as Toolbar } from "./components/Toolbar.svelte";
 export { default as Back } from "./components/Back.svelte";
 export { default as Card } from "./components/Card.svelte";
 export { default as InfiniteScroll } from "./components/InfiniteScroll.svelte";
-export { default as Title } from "./components/Title.svelte";
+export { default as HeaderTitle } from "./components/HeaderTitle.svelte";
 
 export { default as IconArrowRight } from "./icons/IconArrowRight.svelte";
 export { default as IconBackIosNew } from "./icons/IconBackIosNew.svelte";
