+++ src/routes/__layout.svelte
@@ -3,7 +3,7 @@
   import DocsMenu from "$docs/components/DocsMenu.svelte";
   import { page } from "$app/stores";
   import { goto, afterNavigate } from "$app/navigation";
-  import Title from "../lib/components/Title.svelte";
+  import HeaderTitle from "../lib/components/HeaderTitle.svelte";
 
   let navHistory: { from: URL | null; to: URL }[] = [];
   let back = false;
@@ -19,7 +19,7 @@
 </script>
 
 <Layout {back} on:nnsBack={async () => await goBack()}>
-  <Title slot="title">GIX Components</Title>
+  <HeaderTitle slot="title">GIX Components</HeaderTitle>
 
   <DocsMenu slot="menu-items" on:click />
