+++ src/routes/components.svelte
@@ -17,10 +17,10 @@
     <p>Header, menu and footer. The structure of your dapp.</p>
   </Card>
 
-  <Card role="link" on:click={() => goto("/components/title")}>
-    <h2 class="title" slot="start">Layout</h2>
+  <Card role="link" on:click={() => goto("/components/header-title")}>
+    <h2 class="title" slot="start">Header Title</h2>
 
-    <p>An opinionated container to display a title.</p>
+    <p>An opinionated title for the dapp.</p>
   </Card>
 
   <Card role="link" on:click={() => goto("/components/grid")}>
@@ -30,7 +30,7 @@
   </Card>
 
   <Card role="link" on:click={() => goto("/components/infinite-scroll")}>
-    <h2 class="title" slot="start">Infinite scroll</h2>
+    <h2 class="title" slot="start">Infinite Scroll</h2>
 
     <p>Calls an action when the user scrolls a specified distance in a list.</p>
   </Card>
