+++ src/routes/components/header-title.md
@@ -1,9 +1,9 @@
 # Title
 
-An opinionated container to display a title. Commonly used to display the title of the pages within the nav header of the dapp.
+An opinionated container to display a title in the nav header of the dapp.
 
 ```html
-<title>Hello World 👋</title>
+<HeaderTitle>Hello World 👋</HeaderTitle>
 ```
 
 ## Slots
